backend/requirements.txt
uvicorn
strawberry-graphql
fastapi
